requirements.txt
boto3>=1.34.0
botocore>=1.34.0
webvtt-py>=0.4.6
pyyaml>=6.0
click>=8.1.0
python-dotenv>=1.0.0
